backend/form-worker/pyproject.toml
[tool.poetry]
name = "form-worker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dependency-injector = "^4.43.0"
aiologger = "^0.7.0"
aiokafka = "^0.12.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
aiosmtplib = "^3.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
